Добавление констант
