+++ src/constant.js
@@ -1,0 +1,8 @@
+export const POINT_COUNT = 3;
+
+export const RenderPosition = {
+  BEFORE_BEGIN: 'beforebegin',
+  AFTER_BEGIN: 'afterbegin',
+  BEFORE_END: 'beforeend',
+  AFTER_END: 'afterend',
+};
