--- src/helpers/config.js
@@ -1,3 +1,0 @@
-export const config = {
-  POINT_COUNT: 3
-};
