+++ src/helpers/render.js
@@ -1,9 +1,4 @@
-export const RenderPosition = {
-  BEFORE_BEGIN: 'beforebegin',
-  AFTER_BEGIN: 'afterbegin',
-  BEFORE_END: 'beforeend',
-  AFTER_END: 'afterend',
-};
+import { RenderPosition } from '../constant';
 
 export const renderTemplate = (container, template, place = RenderPosition.BEFORE_END) => {
   container.insertAdjacentHTML(place, template);
